Ensure only trusted Copier templates are managed
Since we allow running arbitrary scripts in the container, executing
custom Copier template updates would allow anyone with write access to a
single repository inside the organization to run arbitrary code,
potentially allowing privilege escalation within the org.

This commit disables updates for any Copier templates, with the
exception of the two trusted sources for the template.

We should migrate all extensions to use the `salt-extensions` source,
allowing to restrict this even further.

The order of the packageRules is important, since the last matching one
overrides configuration from previous ones.